Audio, Mouse, Dev button, Sprites
Added better Audio that I liked that fitted with the game, and made sure that more enemys spawned in (ramping) when the music got faster. Made the mouse disapear when certain levels were played to remove its annoyance and changed the helth item sprite to an apple as my feedback suggested. Not long left to go, only the win screen and difficulty sliders.