Fixs
Due to unforeseen consequences when I exported my game the enemies would move to quickly, this was a massive annoyance that plagued me for a long time and when I did the file rework it made it even worse with the enemies even going invisible. I fixed this by recreating whole enemies, putting files in the right place, and setting a max frame rate on my project of 30 FPS. This was because my enemies move on the process(_delta) function that every frame will move them forward on, if a users computer was making the game run at a high FPS then the enemies would move faster and by setting the FPS it caps the speed. This means that overall the fixes include: files, export and internal mechanics